add methods to api routes
