+++ simple-cms/nuxt/server/api/posts/index.get.ts
@@ -1,0 +1,44 @@
+import { z } from 'zod';
+
+const querySchema = z.object({
+	limit: z.coerce.number().min(1).max(100).default(6),
+	page: z.coerce.number().min(1).default(1),
+});
+
+export default defineEventHandler(async (event) => {
+	const query = await getValidatedQuery(event, querySchema.safeParse);
+
+	if (!query.success) {
+		throw createError({ statusCode: 400, message: 'Invalid query parameters' });
+	}
+
+	const { limit, page } = query.data;
+
+	try {
+		const postsPromise = directusServer.request(
+			readItems('posts', {
+				limit,
+				page,
+				sort: ['-published_at'],
+				fields: ['id', 'title', 'description', 'slug', 'image'],
+				filter: { status: { _eq: 'published' } },
+			}),
+		);
+
+		const countPromise = directusServer.request(
+			readItems('posts', {
+				aggregate: { count: '*' },
+				filter: { status: { _eq: 'published' } },
+			}),
+		);
+
+		let [posts, count] = await Promise.all([postsPromise, countPromise]);
+
+		return {
+			posts,
+			count: Number(count[0]?.count) || 0,
+		};
+	} catch {
+		throw createError({ statusCode: 500, message: 'Failed to fetch paginated posts' });
+	}
+});
