+++ simple-cms/nuxt/server/api/site-data.get.ts
