+++ simple-cms/nuxt/server/api/sitemap.get.ts
@@ -1,27 +1,28 @@
-import { directusServer, readItems } from '../utils/directus-server';
 import type { SitemapUrlInput } from '#sitemap/types';
 
 export default defineSitemapEventHandler(async () => {
 	try {
-		const pages = await directusServer.request(
+		const pagesPromise = directusServer.request(
 			readItems('pages', {
 				fields: ['permalink'],
 			}),
 		);
 
-		const posts = await directusServer.request(
+		const postsPromise = directusServer.request(
 			readItems('posts', {
 				filter: { status: { _eq: 'published' } },
 				fields: ['slug'],
 			}),
 		);
 
+		const [pages, posts] = await Promise.all([pagesPromise, postsPromise]);
+
 		const pageUrls = pages.map((page) => ({
 			loc: `/${page.permalink}`,
 		}));
 
 		const postUrls = posts.map((post) => ({
-			loc: `/posts/${post.slug}`,
+			loc: `/blog/${post.slug}`,
 		}));
 
 		return [...pageUrls, ...postUrls] satisfies SitemapUrlInput[];
